requirements.txt
cloudinary==1.44.0
email-validator
fastapi==0.115.13
ffmpeg-python==0.2.0
google-api-python-client==2.170.0
google-auth
google-auth-oauthlib==1.2.2
google-genai
groq==0.28.0
httpx==0.28.1
ipython==8.12.3
motor==3.7.1
openai==1.88.0
passlib[bcrypt]==1.7.4
Pillow==11.2.1
protobuf==6.31.1
pydantic==2.11.7
pymongo==4.13.2
python-dotenv==1.1.0
python-jose[cryptography]==3.5.0
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
httpx
google-auth
google-auth-oauthlib
google-api-python-client
requests==2.32.4
schedule==1.2.2
starlette
together==1.5.16
typing_extensions==4.14.0
uvicorn==0.34.3
imageio_ffmpeg
